rm unneeded include + change uint32_t to size_t

+++ include/stlink/tools/flash.h
@@ -4,8 +4,6 @@
 #include <stdint.h>
 #include <stlink.h>
 
-#include <stlink/chipid.h>
-
 #define DEBUG_LOG_LEVEL 100
 #define STND_LOG_LEVEL  50
 
@@ -22,7 +20,7 @@ struct flash_opts
     int reset;
     int log_level;
     enum flash_format format;
-    uint32_t flash_size;	/* --flash=n[k][m] */
+    size_t flash_size;	/* --flash=n[k][m] */
 };
 
 #define FLASH_OPTS_INITIALIZER {0, NULL, {}, NULL, 0, 0, 0, 0, 0, 0 }
